--- a/src/hotspot/cpu/riscv/stubRoutines_riscv.hpp
+++ b/src/hotspot/cpu/riscv/stubRoutines_riscv.hpp
@@ -153,6 +153,9 @@ class riscv {
   static void set_completed() {
     _completed = true;
   }
+
+private:
+  static juint    _crc_table[];
 };
 
 #endif // CPU_RISCV_STUBROUTINES_RISCV_HPP