@ArsenyBochkarev 2023-12-11
Performance comparison for disabling/enabling Zba on StarFive VisionFive 2 board:

`-XX:-UseZba`:
| Benchmark                                            | (count)  | Mode | Cnt |     Score |  Error |  Units    |
| --------------------------------------------------- | ---------- | ------- | -----  | ---------- | -------- | --------- |
| CRC32.TestCRC32.testCRC32Update |      64    | thrpt   | 12 | 3563.320 | 3.326 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |     128   | thrpt   | 12 | 1928.837 | 2.234 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |     256   | thrpt   | 12 | 1005.273 | 1.953 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |     512   | thrpt   | 12 |  512.550  | 1.718 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |    2048  | thrpt   | 12 |  130.396  | 0.341 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |   16384 |  thrpt  | 12 |   16.319   | 0.073 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |   65536 | thrpt   | 12 |    3.913    | 0.011 | ops/ms |

 
`-XX:+UseZba`:
| Benchmark                                            | (count)   |  Mode | Cnt   |  Score  | Error  | Units     |
| --------------------------------------------------- | ---------- | ------- | -------- | -------- | -------- | ---------- |
| CRC32.TestCRC32.testCRC32Update |      64     | thrpt |  12 |  4206.654  | 0.547 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |     128    | thrpt |  12 | 2308.843 | 3.565 |  ops/ms |
| CRC32.TestCRC32.testCRC32Update |     256    | thrpt |  12 | 1214.727  | 0.305 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |     512    | thrpt |  12 |  623.173  | 0.651 |  ops/ms |
| CRC32.TestCRC32.testCRC32Update |   2048    | thrpt |  12 |  158.965   | 0.376 |   ops/ms |
| CRC32.TestCRC32.testCRC32Update |   16384  | thrpt |  12 |   19.934  | 0.055 |  ops/ms |
| CRC32.TestCRC32.testCRC32Update |   65536  | thrpt |  12 |    4.730     | 0.007 |     ops/ms |