@robehn 2023-12-12
Thanks! (not a review, just an ack)

There are two other version we probably need also, using carry-less-multiplication.
There is a scalar clmul in Zbc and there is a vclmul in vector.