@ArsenyBochkarev 2023-12-21
Performance measurements for C1 (I used the `-XX:TieredStopAtLevel=1` flag) on T-Head:
| Benchmark                      |  (count) |  Mode | Cnt  |   Score  |  Error |  Units |
| -------------------------------- | --------- | ----------- | ---- | --------- | -------- | -------- |
| CRC32.TestCRC32.testCRC32Update   |    64 | thrpt  | 12 | 3618.323 |   8.083 | ops/ms |
| CRC32.TestCRC32.testCRC32Update   |   128 | thrpt |  12 | 2247.195 |  36.867 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |    256 | thrpt |  12 | 1286.744 |   1.927 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |    512 | thrpt  | 12  | 688.159 |   1.158 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |    2048 |  thrpt  | 12 |  182.966 |   0.220 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |   16384 | thrpt  | 12  |  23.285 |   0.031 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |   65536 | thrpt  | 12   |  5.666 |  0.046 | ops/m |