@ArsenyBochkarev 2023-12-21
| Benchmark                      |  (count) |  Mode | Cnt   |  Score |   Error |  Units |
| ------------------------------------------------ | ----------- | ------ | ---- | ---------- |  ---------- | ---------- |
| CRC32.TestCRC32.testCRC32Update |      64 | thrpt  | 12 | 3617.860 | 17.463 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |    128 | thrpt |  12 | 2253.626 |  4.739 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |    256 | thrpt |  12 | 1242.516 | 83.245 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |    512 | thrpt |  12 |  687.339 |  1.712 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |   2048 | thrpt |  12 |  183.016 |  0.258 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |  16384 | thrpt |  12 |   23.368 |  0.133 |  ops/ms |
| CRC32.TestCRC32.testCRC32Update  |  65536 | thrpt  | 12  |   5.640 |  0.023 | ops/ms |