@ArsenyBochkarev 2024-01-09
Hello again everyone! I was able to optimize regressions for most cases on big amount of data by partially unrolling the big loop and disposing from loop counter (previously in `len` register). Results for `-XX:+UseZba` of StarFive VisionFive2 board:

| Benchmark                     |   (count) |  Mode | Cnt  |   Score  |  Error  | Units |
| ------------------------------ | ------------ | --------- | ----- | ---------- | ----------- | ----- |
| CRC32.TestCRC32.testCRC32Update |      64 | thrpt |  12 | 4215.728 | 3.972 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |     128 | thrpt |  12 | 2607.882 |  1.627 |  ops/ms |
| CRC32.TestCRC32.testCRC32Update |     256 | thrpt |  12 | 1364.899 |  8.857 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |     512 | thrpt |  12 |  704.316 |  3.222 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |    2048 | thrpt  | 12 |  180.738 |  0.474 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |   16384 | thrpt  | 12 |   22.722 |  0.059 | ops/ms |
| CRC32.TestCRC32.testCRC32Update |   65536 | thrpt  | 12 |    5.327 |  0.019 | ops/ms |

while the results for `-XX:-UseCRC32Intrinsics` are [here](https://github.com/openjdk/jdk/pull/17046#issuecomment-1850364667)