@ArsenyBochkarev 2024-06-29
Hi all! Can anyone take another look, please?

Current numbers on VisionFive2:
with intrinsic enabled:

| Benchmark                      |  (count)  | Mode | Cnt |    Score  |  Error |  Units |
| -------------------------------------------------- | ------- | ------- | --- | ------------ | ------- | ----------- |
| CRC32.TestCRC32.testCRC32Update  |     64 | thrpt  | 12 | 4644.129 |  9.566 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |    128 | thrpt  | 12 | 2911.927 | 12.866 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |    256 | thrpt  | 12 | 1538.630 |  5.463 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |    512 | thrpt  | 12  | 799.100 |  3.216 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |   2048 | thrpt |  12  | 205.947 |  0.236 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |  16384 | thrpt |  12  |  25.880 |  0.069 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |  65536 | thrpt  | 12 |    6.022 |  0.020 | ops/ms |
| CRC32.TestCRC32.testCRC32Update   | 131072 |  thrpt   | 40 |  3.030 |  0.001 |  ops/ms |
| CRC32.TestCRC32.testCRC32Update   | 262144  | thrpt   | 40 |  1.514 |  0.001 |  ops/ms |
| CRC32.TestCRC32.testCRC32Update  |  524288  | thrpt   | 40 |  0.757 |  0.001 |  ops/ms |
| CRC32.TestCRC32.testCRC32Update |  2097152  | thrpt  |  40 |  0.188 |  0.001 |  ops/ms |

and without intrinsic enabled:
| Benchmark                    |    (count) |  Mode | Cnt  |  Score  |  Error  | Units |
| ------------------------------- | ------------ | --------- | ----- | ---------| ---------- | ------ |
| CRC32.TestCRC32.testCRC32Update  |     64    | thrpt |  12 | 1390.530 | 42.217 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |    128   | thrpt |  12 | 1109.742 | 24.201 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |    256   | thrpt |  12 |  805.345  | 12.155 |  ops/ms |
| CRC32.TestCRC32.testCRC32Update  |    512   | thrpt |  12 |  520.965  |  5.651 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |   2048  | thrpt |  12 |  169.591  |  0.747 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |  16384 | thrpt |  12 |   22.624   |  0.139 | ops/ms |
| CRC32.TestCRC32.testCRC32Update  |  65536 | thrpt |  12 |    5.430    |  0.016 | ops/ms |
| CRC32.TestCRC32.testCRC32Update	| 131072	| thrpt | 40 | 2.729 | 0.003 | ops/ms |
| CRC32.TestCRC32.testCRC32Update	| 262144	| thrpt | 40 | 1.367 | 0.001 | ops/ms |
| CRC32.TestCRC32.testCRC32Update	| 524288	| thrpt | 40 | 0.684 | 0.001 | ops/ms |
| CRC32.TestCRC32.testCRC32Update	| 2097152 | thrpt | 40 | 0.170 | 0.001 | ops/ms |